+++ src/components/Switch/Switch.tsx
@@ -17,18 +17,23 @@ export interface SwitchProps extends Omit<BlueprintSwitchProps, "onChange"> {
      * class names
      */
     className?: string;
+
+    /** Adds the 'nodrag' class to the element, preventing dragging via the Switch element. Default: true */
+    noDrag?: boolean
 }
 
-export const Switch = ({ onChange, className, label, ...otherProps }: SwitchProps) => {
+export const Switch = ({ onChange, className, label, noDrag = true, ...otherProps }: SwitchProps) => {
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         if (onChange) {
             onChange(!!e.target?.checked);
         }
     };
 
+    const noDragClass = noDrag ? "nodrag " : ""
+
     return (
         <BlueprintSwitch
-            className={`${eccgui}-switch ${className ?? ""} ${
+            className={`${eccgui}-switch ${noDragClass}${className ?? ""} ${
                 label && !otherProps.labelElement ? BlueprintClasses.INLINE : ""
             }`}
             labelElement={